+++ mlxtend/evaluate/__init__.py
@@ -8,6 +8,7 @@ from .scoring import scoring
 from .confusion_matrix import confusion_matrix
 from .lift_score import lift_score
 from .mcnemar import mcnemar_table
+from .mcnemar import mcnemar_tables
 from .mcnemar import mcnemar
 from .bootstrap import bootstrap
 from .bootstrap_outofbag import BootstrapOutOfBag
@@ -16,6 +17,7 @@ from .permutation import permutation_test
 
 
 __all__ = ["scoring", "confusion_matrix",
-           "mcnemar_table", "mcnemar", "lift_score",
+           "mcnemar_table", "mcnemar_tables",
+           "mcnemar", "lift_score",
            "bootstrap", "permutation_test",
            "BootstrapOutOfBag", "bootstrap_point632_score"]